Modularize and cleanup
